+++ src/main/java/com/karinasasaki/gerenciador_de_tarefas/repositories/TarefaRepository.java
@@ -3,5 +3,5 @@ package com.karinasasaki.gerenciador_de_tarefas.repositories;
 import com.karinasasaki.gerenciador_de_tarefas.entities.Tarefa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TarefaRepository extends JpaRepository<Tarefa, Long> {
+public interface TarefaRepository extends JpaRepository<Tarefa, Integer> {
 }
